Browser login

+++ app/src/main/java/com/yandex/yaloginsdk/Token.java
@@ -14,9 +14,9 @@ public abstract class Token implements Parcelable {
     @NonNull
     public abstract String type();
 
-    public abstract double expiresIn();
+    public abstract long expiresIn();
 
-    public static Token create(@NonNull String token, @NonNull String type, double expiresIn) {
+    public static Token create(@NonNull String token, @NonNull String type, long expiresIn) {
         return new AutoValue_Token(token, type, expiresIn);
     }
 }
